requirements.txt
basicsr>=1.4.2
facexlib>=0.2.5
lmdb
numpy~=1.24.4
opencv-contrib-python~=4.8.0.74
pyyaml~=6.0.1
scipy
tb-nightly
torch>=1.7
torchvision~=0.16.0
tqdm~=4.66.1
yapf

matplotlib~=3.7.1
pillow~=9.4.0
setuptools~=73.0.1
uvicorn~=0.30.3
fastapi~=0.111.1
scikit-image~=0.24.0
face_recognition